pyproject.toml
[tool.poetry]
name = "surton-ip-tools"
version = "0.1.0"
description = ""
authors = ["Surton <devs@surton.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
dnspython = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
